+++ UCR.Core/Utilities/AxisHelpers/DeadZoneHelper.cs
@@ -43,9 +43,9 @@ namespace HidWizards.UCR.Core.Utilities.AxisHelpers
             }
         }
 
-        public long ApplyRangeDeadZone(long value)
+        public short ApplyRangeDeadZone(short value)
         {
-            var absValue = Math.Abs(value);
+            var absValue = Math.Abs((int)value);
             if (absValue < Math.Round(_deadzoneCutoff))
             {
                 return 0;
@@ -53,10 +53,10 @@ namespace HidWizards.UCR.Core.Utilities.AxisHelpers
 
             var sign = Math.Sign(value);
             var adjustedValue = (absValue - _deadzoneCutoff) * _scaleFactor;
-            var newValue = (long) Math.Round(adjustedValue * sign);
+            var newValue = (int) Math.Round(adjustedValue * sign);
             if (newValue < -32768) newValue = -32768;   // ToDo: Negative values can go up to -32777 (9 over), can this be improved?
             //Debug.WriteLine($"Pre-DZ: {value}, Post-DZ: {newValue}, Cutoff: {_deadzoneCutoff}");
-            return newValue;
+            return (short) newValue;
         }
     }
 }
